moved redo query to util function
